+++ src/Stormancer.Plugins/Steam/Stormancer.Server.Plugins.Steam/SteamPartyEventHandler.cs
@@ -224,6 +224,8 @@ namespace Stormancer.Server.Plugins.Steam
                                     partySettingsDto.PublicServerData["SteamIDLobby"] = steamIDLobby.ToString();
                                     _ = ctx.Party.UpdateSettings(partySettingsDto, CancellationToken.None);
                                     data.SteamIDLobby = steamIDLobby;
+                                    data.UserData[ctx.Session.SessionId] = new SteamUserData { SessionId = ctx.Session.SessionId, SteamId = (ulong)steamId };
+                                    data.IncrementNumMembers();
                                     data.IsJoinable = joinable;
                                 }
                                 else
@@ -329,6 +331,11 @@ namespace Stormancer.Server.Plugins.Steam
                         await _steamService.RemoveUserFromLobby(steamUserData.SteamId, data.SteamIDLobby);
                         data.DecrementNumMembers();
                         data.UserData.TryRemove(sessionId, out var _);
+
+                        if(data.UserData.Count == 0)
+                        {
+                            data.SteamIDLobby = 0;
+                        }
                     });
                 }
             }