+++ src/Stormancer.Plugins/Party/Stormancer.Server.Plugins.Party/Changelog.rst
@@ -13,12 +13,12 @@ Unreleased
 ----------
 Changed
 *******
-- Remove Dependency to Stormancer.Abstractions.Server because it's implicit.
 - Send IndexedDocument as part of party updated events.
 - Add configuration and event handler extensibility points to control when ready state is reset, validate ready state changes and user data updates.
 - Parties are now destroyed immediatly after the last player left to prevent players from joining afterwards.
 - UpdateSettings method now takes a PartyState object as lambda argument.
 - PartyUpdateDto new takes a PartyState object as lambda argument.
+- Automatically destroy scene when last player disconnects.
 
 Added
 *****